change type double to float

+++ EFCoreWithPostgreSQL/Models/OrderWithOrderDetailJson/OrderWithOrderDetailEntity.cs
@@ -11,8 +11,8 @@ namespace EFCoreJsonApp.Models.OrderWithOrderDetail
     public class OrderDetailsJson
     {
         public string ItemName { get; set; }
-        public double Price { get; set; }
+        public float Price { get; set; }
         public int Quantity { get; set; }
-        public double Total { get; set; }
+        public float Total { get; set; }
     }
 }
